Consolidate settings logic into one file
