+++ src/actions/SettingsActions.tsx
@@ -1,20 +1,16 @@
+import { asArray, asBoolean, asMaybe, asNumber, asObject, asOptional, asString, asValue, Cleaner } from 'cleaners'
 import { EdgeAccount, EdgeDenomination, EdgeSwapPluginType } from 'edge-core-js'
 import { disableTouchId, enableTouchId } from 'edge-login-ui-rn'
 import * as React from 'react'
 
 import { ButtonsModal } from '../components/modals/ButtonsModal'
+import { asSortOption, SortOption } from '../components/modals/WalletListSortModal'
 import { Airship, showError } from '../components/services/AirshipInstance'
 import { lstrings } from '../locales/strings'
-import {
-  setAutoLogoutTimeInSecondsRequest as setAutoLogoutTimeInSecondsRequestAccountSettings,
-  setDefaultFiatRequest as setDefaultFiatRequestAccountSettings,
-  setDenominationKeyRequest as setDenominationKeyRequestAccountSettings,
-  setPreferredSwapPluginId as setPreferredSwapPluginIdAccountSettings,
-  setPreferredSwapPluginType as setPreferredSwapPluginTypeAccountSettings
-} from '../modules/Core/Account/settings'
 import { convertCurrency } from '../selectors/WalletSelectors'
 import { ThunkAction } from '../types/reduxTypes'
 import { NavigationBase } from '../types/routerTypes'
+import { asMostRecentWallet, MostRecentWallet } from '../types/types'
 import { logActivity } from '../util/logger'
 import { DECIMAL_PRECISION } from '../util/utils'
 import { validatePassword } from './AccountActions'
@@ -41,7 +37,7 @@ export function setAutoLogoutTimeInSecondsRequest(autoLogoutTimeInSeconds: numbe
   return async (dispatch, getState) => {
     const state = getState()
     const { account } = state.core
-    await setAutoLogoutTimeInSecondsRequestAccountSettings(account, autoLogoutTimeInSeconds)
+    await writeAutoLogoutTimeInSeconds(account, autoLogoutTimeInSeconds)
     dispatch({
       type: 'UI/SETTINGS/SET_AUTO_LOGOUT_TIME',
       data: { autoLogoutTimeInSeconds }
@@ -61,7 +57,7 @@ export function setDefaultFiatRequest(defaultFiat: string): ThunkAction<Promise<
     const previousDefaultIsoFiat = state.ui.settings.defaultIsoFiat
 
     // update default fiat in account settings
-    await setDefaultFiatRequestAccountSettings(account, defaultFiat)
+    await writeDefaultFiatSetting(account, defaultFiat)
 
     // update default fiat in settings
     dispatch({
@@ -96,7 +92,7 @@ export function setPreferredSwapPluginId(pluginId: string | undefined): ThunkAct
   return (dispatch, getState) => {
     const state = getState()
     const { account } = state.core
-    setPreferredSwapPluginIdAccountSettings(account, pluginId)
+    writePreferredSwapPluginId(account, pluginId)
       .then(() => {
         dispatch({
           type: 'UI/SETTINGS/SET_PREFERRED_SWAP_PLUGIN',
@@ -115,7 +111,7 @@ export function setPreferredSwapPluginType(swapPluginType: EdgeSwapPluginType | 
   return (dispatch, getState) => {
     const state = getState()
     const { account } = state.core
-    setPreferredSwapPluginTypeAccountSettings(account, swapPluginType)
+    writePreferredSwapPluginType(account, swapPluginType)
       .then(() => {
         dispatch({
           type: 'UI/SETTINGS/SET_PREFERRED_SWAP_PLUGIN_TYPE',
@@ -136,7 +132,7 @@ export function setDenominationKeyRequest(pluginId: string, currencyCode: string
     const state = getState()
     const { account } = state.core
 
-    return await setDenominationKeyRequestAccountSettings(account, pluginId, currencyCode, denomination)
+    return await writeDenominationKeySetting(account, pluginId, currencyCode, denomination)
       .then(() =>
         dispatch({
           type: 'UI/SETTINGS/SET_DENOMINATION_KEY',
@@ -263,4 +259,161 @@ export function showRestoreWalletsModal(navigation: NavigationBase): ThunkAction
       navigation.navigate('walletsTab', { screen: 'walletList' })
     }
   }
+}
+
+export const PASSWORD_RECOVERY_REMINDERS_SHOWN = {
+  '20': false,
+  '200': false,
+  '2000': false,
+  '20000': false,
+  '200000': false
+}
+
+export type PasswordReminderTime = keyof typeof PASSWORD_RECOVERY_REMINDERS_SHOWN
+
+export const asCurrencyCodeDenom = asObject({
+  name: asString,
+  multiplier: asString,
+  symbol: asOptional(asString)
+})
+
+const asDenominationSettings = asObject(asOptional(asObject(asMaybe(asCurrencyCodeDenom))))
+
+export type DenominationSettings = ReturnType<typeof asDenominationSettings>
+export const asSwapPluginType: Cleaner<'CEX' | 'DEX'> = asValue('CEX', 'DEX')
+
+export type SecurityCheckedWallets = Record<string, { checked: boolean; modalShown: number }>
+
+const asSecurityCheckedWallets: Cleaner<SecurityCheckedWallets> = asObject(
+  asObject({
+    checked: asBoolean,
+    modalShown: asNumber
+  })
+)
+
+export const asSyncedAccountSettings = asObject({
+  autoLogoutTimeInSeconds: asOptional(asNumber, 3600),
+  defaultFiat: asOptional(asString, 'USD'),
+  defaultIsoFiat: asOptional(asString, 'iso:USD'),
+  preferredSwapPluginId: asOptional(asString, ''),
+  preferredSwapPluginType: asOptional(asSwapPluginType),
+  countryCode: asOptional(asString, ''),
+  mostRecentWallets: asOptional(asArray(asMostRecentWallet), () => []),
+  passwordRecoveryRemindersShown: asOptional(
+    asObject({
+      '20': asBoolean,
+      '200': asBoolean,
+      '2000': asBoolean,
+      '20000': asBoolean,
+      '200000': asBoolean
+    }),
+    PASSWORD_RECOVERY_REMINDERS_SHOWN
+  ),
+  walletsSort: asOptional(asSortOption, 'manual'),
+  denominationSettings: asOptional<DenominationSettings>(asDenominationSettings, () => ({})),
+  securityCheckedWallets: asMaybe<SecurityCheckedWallets>(asSecurityCheckedWallets, () => ({}))
+})
+
+// Default Account Settings
+export const SYNCED_ACCOUNT_DEFAULTS = asSyncedAccountSettings({})
+
+const SYNCED_SETTINGS_FILENAME = 'Settings.json'
+
+// Account Settings
+const writeAutoLogoutTimeInSeconds = async (account: EdgeAccount, autoLogoutTimeInSeconds: number) =>
+  await readSyncedSettings(account).then(async settings => {
+    const updatedSettings = { ...settings, autoLogoutTimeInSeconds }
+    return await writeSyncedSettings(account, updatedSettings)
+  })
+
+const writeDefaultFiatSetting = async (account: EdgeAccount, defaultFiat: string) =>
+  await readSyncedSettings(account).then(async settings => {
+    const updatedSettings = {
+      ...settings,
+      defaultFiat,
+      defaultIsoFiat: `iso:${defaultFiat}`
+    }
+    return await writeSyncedSettings(account, updatedSettings)
+  })
+
+const writePreferredSwapPluginId = async (account: EdgeAccount, pluginId: string | undefined) => {
+  return await readSyncedSettings(account).then(async settings => {
+    const updatedSettings = {
+      ...settings,
+      preferredSwapPluginId: pluginId == null ? '' : pluginId,
+      preferredSwapPluginType: undefined
+    }
+    return await writeSyncedSettings(account, updatedSettings)
+  })
+}
+
+const writePreferredSwapPluginType = async (account: EdgeAccount, swapPluginType: EdgeSwapPluginType | undefined) => {
+  return await readSyncedSettings(account).then(async settings => {
+    const updatedSettings = {
+      ...settings,
+      preferredSwapPluginType: swapPluginType,
+      preferredSwapPluginId: ''
+    }
+    return await writeSyncedSettings(account, updatedSettings)
+  })
+}
+
+export const writeMostRecentWalletsSelected = async (account: EdgeAccount, mostRecentWallets: MostRecentWallet[]) =>
+  await readSyncedSettings(account).then(async settings => {
+    const updatedSettings = { ...settings, mostRecentWallets }
+    return await writeSyncedSettings(account, updatedSettings)
+  })
+
+export const writeWalletsSort = async (account: EdgeAccount, walletsSort: SortOption) =>
+  await readSyncedSettings(account).then(async settings => {
+    const updatedSettings = { ...settings, walletsSort }
+    return await writeSyncedSettings(account, updatedSettings)
+  })
+
+export async function writePasswordRecoveryReminders(account: EdgeAccount, level: PasswordReminderTime) {
+  const settings = await readSyncedSettings(account)
+  const passwordRecoveryRemindersShown = {
+    ...settings.passwordRecoveryRemindersShown
+  }
+  passwordRecoveryRemindersShown[level] = true
+  const updatedSettings = { ...settings, passwordRecoveryRemindersShown }
+  return await writeSyncedSettings(account, updatedSettings)
+}
+
+// Currency Settings
+const writeDenominationKeySetting = async (account: EdgeAccount, pluginId: string, currencyCode: string, denomination: EdgeDenomination) =>
+  await readSyncedSettings(account).then(async settings => {
+    const updatedSettings = updateCurrencySettings(settings, pluginId, currencyCode, denomination)
+    return await writeSyncedSettings(account, updatedSettings)
+  })
+
+// Helper Functions
+export async function readSyncedSettings(account: EdgeAccount): Promise<ReturnType<typeof asSyncedAccountSettings>> {
+  try {
+    if (account?.disklet?.getText == null) return SYNCED_ACCOUNT_DEFAULTS
+    const text = await account.disklet.getText(SYNCED_SETTINGS_FILENAME)
+    const settingsFromFile = JSON.parse(text)
+    return asSyncedAccountSettings(settingsFromFile)
+  } catch (e: any) {
+    console.log(e)
+    // If Settings.json doesn't exist yet, create it, and return it
+    await writeSyncedSettings(account, SYNCED_ACCOUNT_DEFAULTS)
+    return SYNCED_ACCOUNT_DEFAULTS
+  }
+}
+
+export async function writeSyncedSettings(account: EdgeAccount, settings: object): Promise<void> {
+  const text = JSON.stringify(settings)
+  if (account?.disklet?.setText == null) return
+  await account.disklet.setText(SYNCED_SETTINGS_FILENAME, text)
+}
+
+const updateCurrencySettings = (currentSettings: any, pluginId: string, currencyCode: string, denomination: EdgeDenomination) => {
+  // update with new settings
+  const updatedSettings = {
+    ...currentSettings
+  }
+  if (updatedSettings.denominationSettings[pluginId] == null) updatedSettings.denominationSettings[pluginId] = {}
+  updatedSettings.denominationSettings[pluginId][currencyCode] = denomination
+  return updatedSettings
 }
