+++ src/actions/WalletListActions.tsx
@@ -2,12 +2,12 @@ import * as React from 'react'
 import { Linking } from 'react-native'
 import { sprintf } from 'sprintf-js'
 
+import { writeWalletsSort } from '../actions/SettingsActions'
 import { showFullScreenSpinner } from '../components/modals/AirshipFullScreenSpinner'
 import { ButtonsModal } from '../components/modals/ButtonsModal'
 import { SortOption } from '../components/modals/WalletListSortModal'
 import { Airship, showError } from '../components/services/AirshipInstance'
 import { lstrings } from '../locales/strings'
-import { setWalletsSort } from '../modules/Core/Account/settings'
 import { GetState, ThunkAction } from '../types/reduxTypes'
 import { NavigationBase } from '../types/routerTypes'
 import { getCreateWalletType } from '../util/CurrencyInfoHelpers'
@@ -24,7 +24,7 @@ export function updateWalletsSort(walletsSort: SortOption): ThunkAction<void> {
       type: 'UI/SETTINGS/SET_WALLETS_SORT',
       data: { walletsSort }
     })
-    setWalletsSort(account, walletsSort).catch(showError)
+    writeWalletsSort(account, walletsSort).catch(showError)
   }
 }
